fix: address PR #112 CodeRabbit review (round 2)
- Split AIActionAuditStore to separate file (500-line limit)
- Add tenantId filter and entry immutability to audit store
- Fix limit=0 handling in audit query
- Don't allocate limiter state on read paths
- Use null-byte separator for collision-safe identity keys
- Reorder masking rules (credit card before phone)
- Handle BigInt/circular in JSON.stringify masking
